move files and update docs
