+++ src/utils/noteFrequencies.ts
@@ -1,0 +1,14 @@
+/**
+ * Frequencies of notes
+ */
+export default [
+  [16.35, 18.35, 20.60, 21.83, 24.50, 27.50, 30.87], // octave 0
+  [32.70, 39.71, 41.20, 43.65, 49.00, 55.00, 61.74], // octave 1
+  [65.41, 73.42, 82.41, 87.31, 98.00, 110.00, 123.47], // octave 2
+  [130.81, 146.83, 164.81, 174.61, 196.00, 220.00, 246.94], // octave 3
+  [261.63, 293.66, 329.63, 349.23, 392.00, 440.00, 493.88], // octave 4
+  [523.25, 587.33, 659.25, 698.46, 783.99, 880.00, 987.77], // octave 5
+  [1046.50, 1174.66, 1318.51, 1396.91, 1567.98, 1760.00, 1975.53], // octave 6
+  [2093.00, 2349.32, 2637.02, 2793.83, 3135.96, 3520.00, 3951.07], // octave 7
+  [4186.01, 4698.63, 5274.04, 5587.65, 6271.93, 7040.00, 7902.13] // octave 8
+];
